Reorganize TraitsExecutor tests
